% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corstars.R
\name{corstars}
\alias{corstars}
\title{Correlation Matrix with Stars}
\usage{
corstars(x, method = c("pearson", "spearman"),
  removeTriangle = c("upper", "lower"), result = c("none", "html",
  "latex"))
}
\arguments{
\item{x, }{}

\item{method=c("pearson", }{"spearman")}

\item{removeTriangle=c("upper", }{"lower")}

\item{result=c("none", }{"html", "latex")}
}
\description{
Creates a correlation matrix in html file or latext.
}
\examples{
install.packages("xtable")
library(xtable)
corstars()
x <- print(xtable(corstars(my_data)), type="html")
dput(x, file = "output.html")
browseURL("output.html")
}
\keyword{Correlation,}
\keyword{Pearson's}
\keyword{R.}
