% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Load Function}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The library path.}

\item{pkgname}{The package name.}
}
\description{
Automatically installs and loads required packages when the package is loaded into the namespace.
}
