% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{install_load_packages}
\alias{install_load_packages}
\title{Install and Load Required Packages}
\arguments{
\item{packages}{A character vector of package names to check, install, and load.}
}
\description{
This code checks for the installation of required packages and installs any missing ones.
It then loads all the specified packages into the R session.
}
\details{
If a package is not installed, it will be installed with all its dependencies.
Messages will indicate any missing packages that are being installed.
}
\examples{
packages <- c("ggplot2", "dplyr")
for (package in packages) {
  if (!require(package, character.only = TRUE)) {
    install.packages(package, dependencies = TRUE)
  }
  library(package, character.only = TRUE)
}
}
