% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_outlier.R
\name{nick_remOutliers}
\alias{nick_remOutliers}
\title{Remove Outliers from Data}
\usage{
nick_remOutliers(
  data,
  variable,
  group1 = NULL,
  group2 = NULL,
  remove_outliers = TRUE
)
}
\arguments{
\item{data}{A dataframe from which outliers should be removed.}

\item{variable}{The name of the variable to clean (as a string).}

\item{group1}{An optional first grouping variable (as a string).}

\item{group2}{An optional second grouping variable (as a string).}

\item{remove_outliers}{Logical; if TRUE, outliers are removed (Default: TRUE).}
}
\value{
A cleaned dataframe without outliers.
}
\description{
This function removes outliers based on the Interquartile Range (IQR).
Optionally, grouping variables can be specified to calculate outliers within each group.
}
\examples{
df <- data.frame(values = c(1, 2, 3, 100, 5), group1 = c("A", "A", "B", "B", "B"), group2 = c("X", "X", "Y", "Y", "Y"))
remove_outliers(df, variable = "values")
remove_outliers(df, variable = "values", group1 = "group1")
remove_outliers(df, variable = "values", group1 = "group1", group2 = "group2")
}
