% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{A variable}
}
\description{
This function allows you normalise a variable. The function takes one variable as an input, and then subtracts the mean and divides by the standard deviation of the variable.
}
\examples{
normalize()
}
\keyword{Centering}
\keyword{Normalisation,}
\keyword{Standardisation,}
