% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_outlier.R
\name{nick_remOutliers}
\alias{nick_remOutliers}
\title{Entferne Ausreißer aus Daten}
\usage{
nick_remOutliers(
  data,
  variable,
  group1 = NULL,
  group2 = NULL,
  remove_outliers = TRUE
)
}
\arguments{
\item{data}{Ein Dataframe, aus dem Ausreißer entfernt werden sollen.}

\item{variable}{Der Name der zu bereinigenden Variable (als String).}

\item{group1}{Eine optionale erste Gruppierungsvariable (als String).}

\item{group2}{Eine optionale zweite Gruppierungsvariable (als String).}

\item{remove_outliers}{Logical; wenn TRUE, werden Ausreißer entfernt (Default: TRUE).}
}
\value{
Ein bereinigter Dataframe ohne Ausreißer.
}
\description{
Diese Funktion entfernt Ausreißer basierend auf dem Interquartilsabstand (IQR).
Optional können Gruppierungsvariablen angegeben werden, um die Ausreißer innerhalb jeder Gruppe zu berechnen.
}
\examples{
df <- data.frame(values = c(1, 2, 3, 100, 5), group1 = c("A", "A", "B", "B", "B"), group2 = c("X", "X", "Y", "Y", "Y"))
remove_outliers(df, variable = "values")
remove_outliers(df, variable = "values", group1 = "group1")
remove_outliers(df, variable = "values", group1 = "group1", group2 = "group2")
}
